requirements.txt
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
executing==1.2.0
h11==0.14.0
idna==3.4
ipython==8.15.0
jedi==0.19.0
matplotlib-inline==0.1.6
outcome==1.3.0.post0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
traitlets==5.10.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wcwidth==0.2.6
wsproto==1.2.0
